mcp-servers/jira-mcp/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jira==3.8.0
google-cloud-secret-manager==2.20.0
requests==2.32.3